condition_api/requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow
opencv-python
python-multipart
